chore: refactor cli
